refactor: move TransactionDirection to the tari_common_types crate
