+++ applications/tari_console_wallet/src/ui/components/transactions_tab.rs
@@ -7,8 +7,7 @@ use crate::ui::{
     MAX_WIDTH,
 };
 use chrono::{DateTime, Local};
-use tari_common_types::transaction::TransactionStatus;
-use tari_wallet::transaction_service::storage::models::TransactionDirection;
+use tari_common_types::transaction::{TransactionDirection, TransactionStatus};
 use tokio::runtime::Handle;
 use tui::{
     backend::Backend,
