+++ base_layer/common_types/src/transaction.rs
@@ -69,3 +69,40 @@ impl Display for TransactionStatus {
         }
     }
 }
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub enum TransactionDirection {
+    Inbound,
+    Outbound,
+    Unknown,
+}
+
+#[derive(Debug, Error)]
+#[error("Invalid TransactionDirection: {code}")]
+pub struct TransactionDirectionError {
+    pub code: i32,
+}
+
+impl TryFrom<i32> for TransactionDirection {
+    type Error = TransactionDirectionError;
+
+    fn try_from(value: i32) -> Result<Self, Self::Error> {
+        match value {
+            0 => Ok(TransactionDirection::Inbound),
+            1 => Ok(TransactionDirection::Outbound),
+            2 => Ok(TransactionDirection::Unknown),
+            code => Err(TransactionDirectionError { code }),
+        }
+    }
+}
+
+impl Display for TransactionDirection {
+    fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), Error> {
+        // No struct or tuple variants
+        match self {
+            TransactionDirection::Inbound => write!(f, "Inbound"),
+            TransactionDirection::Outbound => write!(f, "Outbound"),
+            TransactionDirection::Unknown => write!(f, "Unknown"),
+        }
+    }
+}
